pyproject.toml
[project]
name = "trip-planner"
version = "1.0.0"
description = "AI-powered multi-agent trip planning assistant"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Trip Planner Team" }
]
keywords = ["ai", "travel", "planning", "agents", "google-adk"]

dependencies = [
    "google-adk>=0.1.0",
    "google-generativeai>=0.3.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "opik>=0.1.0",
    "httpx>=0.25.0",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
trip-planner = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]
